requirements.txt
msedge_selenium_tools==3.141.3
selenium==3.141.0
